добавлены методы, принимающие на вход json-строку

+++ src/main/java/com/core/Groups.java
@@ -39,71 +39,93 @@ public class Groups {
     @GET
     @Path("{id: \\d+}")
     @Produces(MediaType.APPLICATION_JSON)
-    public String getGroupById(@PathParam("id") int id) {
+    public Response getGroupById(@PathParam("id") int id) {
         GroupService groupService = new GroupService();
         Group group = groupService.getGroupById(id);
         if (group == null) {
             throw new WebApplicationException(Response.Status.NOT_FOUND);
         }
         groupService.stop();
-        return "{\"group\": " + new Gson().toJson(group) + "}";
+        return Response.ok(group, MediaType.APPLICATION_JSON).build();
     }
 
     @GET
     @Path("{name}")
     @Produces(MediaType.APPLICATION_JSON)
-    public String getGroupByName(@PathParam("name") String name) {
+    public Response getGroupByName(@PathParam("name") String name) {
         GroupService groupService = new GroupService();
         Group group = groupService.getGroupByName(name);
         if (group == null) {
             throw new WebApplicationException(Response.Status.NOT_FOUND);
         }
         groupService.stop();
-        return "{\"group\": " + new Gson().toJson(group) + "}";
+        return Response.ok(group, MediaType.APPLICATION_JSON).build();
     }
 
     @POST
-    @Path("/add")
+    @Path("/add_plain")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response addGroup(@QueryParam("name") String name, @QueryParam("type") String type) {
+    public Response addGroupPlain(@QueryParam("name") String name, @QueryParam("type") String type) {
         GroupService groupService = new GroupService();
         int id = groupService.addGroup(name, type);
         groupService.stop();
         return (id == -1 ? Response.notModified().build() : Response.created(URI.create("/users/" + id)).build());
     }
 
+    @POST
+    @Path("/add")
+    @Produces(MediaType.APPLICATION_JSON)
+    public Response addGroup(String jsonGroup) {
+        Gson gson = new Gson();
+        Group newGroup = gson.fromJson(jsonGroup, Group.class);
+        int id;
+        Group group = null;
+        GroupService groupService = new GroupService();
+        try {
+            id = groupService.addGroup(newGroup.getGroupName(), newGroup.getGroupType());
+            if (id != -1)
+                group = groupService.getGroupById(id);
+        } catch (Exception e) {
+            throw new WebApplicationException(e);
+        } finally {
+            groupService.stop();
+        }
+        return (id == -1 ? Response.notModified().build() : Response.created(URI.create("/" + id)).entity(group).build());
+    }
+
+
     @GET
     @Path("{id: \\d+}/users")
     @Produces(MediaType.APPLICATION_JSON)
-    public String getGroupUsers(@PathParam("id") int id) {
+    public Response getGroupUsers(@PathParam("id") int id) {
         GroupService groupService = new GroupService();
-        UserService userService = new UserService();
-        if (!groupService.checkGroupExists(id))
+        if (!groupService.checkGroupExists(id)) {
+            groupService.stop();
             throw new WebApplicationException(Response.Status.NOT_FOUND);
+        }
         List<User> users = groupService.getGroupUsers(id);
         if (users == null || users.isEmpty()) {
+            groupService.stop();
             throw new WebApplicationException(Response.Status.NOT_FOUND);
         }
-        String response = "{\"users\": [";
-        int i = 0;
-        for (User user : users) {
-            i++;
-            response = (i == users.size() ? response.concat(user.toString()) : response.concat(user.toString() + ", "));
-        }
-        response = response.concat("]}");
-        groupService.stop();
-        userService.stop();
-        return response;
+        return Response.ok(users, MediaType.APPLICATION_JSON).build();
     }
 
     @POST
     @Path("{name}/join")
     @Produces(MediaType.APPLICATION_JSON)
     public Response addGroupUser(@PathParam("name") String name, @QueryParam("user") String login) {
+        UserService userService = new UserService();
         GroupService groupService = new GroupService();
+        if (!userService.checkUserExists(login) || !groupService.checkGroupExists(name)) {
+            userService.stop();
+            groupService.stop();
+            throw new WebApplicationException(Response.Status.NOT_FOUND);
+        }
+        userService.stop();
         int id = groupService.addGroupUser(login, name);
         groupService.stop();
-        return (id == -1 ? Response.notModified().build() : Response.created(URI.create("/users/" + id)).build());
+        return (id == -1 ? Response.notModified().build() : Response.ok().build());
     }
 
     @DELETE
@@ -111,6 +133,10 @@ public class Groups {
     @Produces(MediaType.APPLICATION_JSON)
     public Response delete(@PathParam("id") int id) {
         GroupService groupService = new GroupService();
+        if (!groupService.checkGroupExists(id)) {
+            groupService.stop();
+            throw new WebApplicationException(Response.Status.NOT_FOUND);
+        }
         boolean deleted = groupService.deleteGroup(groupService.getGroupById(id));
         groupService.stop();
         if (!deleted)
@@ -136,20 +162,6 @@ public class Groups {
             response = (i == users.size() ? response.concat(user.toString()) : response.concat(user.toString() + ", "));
         }
         response = response.concat("]}");
-        return response;
-    }
-
-    //todo: убрать параметры, вставить json и добавить парс json'а в объект группы
-    @POST
-    @Path("/add")
-    @Consumes(MediaType.APPLICATION_JSON)
-    @Produces(MediaType.APPLICATION_JSON)
-    public Response addGroup(String jsonGroup) {
-        Gson gson = new Gson();
-        Group group = gson.fromJson(jsonGroup, Group.class);
-        GroupService groupService = new GroupService();
-        int id = groupService.addGroup(group.getGroupName(), group.getGroupType());
-        Response response = (id==-1 ? Response.notModified().build() : Response.created(URI.create("/users/" + id)).build());
         return response;
     }*/
 }
